# src/swagger/3.0.0-rc.1.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Consents - Open Finance Brasil",
...
    "version": "3.0.0-rc.1",
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0"
    },
    "contact": {
      "name": "Governança do Open Finance Brasil - Especificações",
      "email": "gt-interfaces@openbankingbr.org",
      "url": "https://openbanking-brasil.github.io/areadesenvolvedor/"
    }
  },
  "servers": [
    {
      "url": "https://api.banco.com.br/open-banking/consents/v3",
      "description": "Servidor de Produção"
    },
    {
      "url": "https://apih.banco.com.br/open-banking/consents/v3",
      "description": "Servidor de Homologação"
    }
  ],
...
  "paths": {
    "/consents": {
      "post": {
        "tags": [
          "Consents"
        ],
        "summary": "Criar novo pedido de consentimento.",
        "operationId": "consentsPostConsents",
        "description": "Método para a criação de um novo consentimento.",
        "parameters": [
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/xFapiAuthDate"
          },
          {
            "$ref": "#/components/parameters/xFapiCustomerIpAddress"
          },
          {
            "$ref": "#/components/parameters/xFapiInteractionId"
          },
          {
            "$ref": "#/components/parameters/xCustomerUserAgent"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsent"
              }
            }
          },
          "description": "Payload para criação do consentimento.",
          "required": true
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/201ConsentsCreated"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntity"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          },
          "529": {
            "$ref": "#/components/responses/SiteIsOverloaded"
          },
          "default": {
            "description": "Erro inesperado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "consents"
            ]
          }
        ]
      }
    },
    "/consents/{consentId}": {
      "get": {
        "tags": [
          "Consents"
        ],
        "summary": "Obter detalhes do consentimento identificado por consentId.",
        "operationId": "consentsGetConsentsConsentId",
        "description": "Método para obter detalhes do consentimento identificado por consentId.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ConsentId"
          },
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/xFapiAuthDate"
          },
          {
            "$ref": "#/components/parameters/xFapiCustomerIpAddress"
          },
          {
            "$ref": "#/components/parameters/xFapiInteractionId"
          },
          {
            "$ref": "#/components/parameters/xCustomerUserAgent"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200ConsentsConsentIdRead"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          },
          "529": {
            "$ref": "#/components/responses/SiteIsOverloaded"
          },
          "default": {
            "description": "Erro inesperado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "consents"
            ]
          }
        ]
      },
      "delete": {
        "tags": [
          "Consents"
        ],
        "summary": "Deletar / Revogar o consentimento identificado por consentId.",
        "operationId": "consentsDeleteConsentsConsentId",
        "description": "Método para deletar / revogar o consentimento identificado por consentId.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ConsentId"
          },
          {
            "$ref": "#/components/parameters/Authorization"
          },
          {
            "$ref": "#/components/parameters/xFapiAuthDate"
          },
          {
            "$ref": "#/components/parameters/xFapiCustomerIpAddress"
          },
          {
            "$ref": "#/components/parameters/xFapiInteractionId"
          },
          {
            "$ref": "#/components/parameters/xCustomerUserAgent"
          }
        ],
        "responses": {
          "204": {
            "$ref": "#/components/responses/204ConsentsConsentIdDeleted"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntityDelete"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          },
          "529": {
            "$ref": "#/components/responses/SiteIsOverloaded"
          },
          "default": {
            "description": "Erro inesperado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "consents"
            ]
          }
        ]
      }
    },
    "/consents/{consentId}/extensions": {
      "get": {
        "tags": [
          "Consents"
        ],
        "summary": "Obter detalhes de extensões feitas no consentimento identificado por consentId.",
        "operationId": "consentsGetConsentsConsentIdExtensions",
        "description": "Método para obter histórico de extensões de consentimento identificado por consentId.\n\nIMPORTANTE: A lista do payload de resposta deve ser entregue em ordem decrescente pela data de requisição (`data[].requestDateTime`). \nDessa forma, o primeiro item da lista apresentará a mesma data de expiração do consentimento vigente, pois foi a última renovação feita.\n",
        "parameters": [
          {
            "$ref": "#/components/parameters/ConsentId"
          },
          {
            "$ref": "#/components/parameters/AuthorizationExtensions"
          },
          {
            "$ref": "#/components/parameters/xFapiAuthDate"
          },
          {
            "$ref": "#/components/parameters/xFapiCustomerIpAddress"
          },
          {
            "$ref": "#/components/parameters/xFapiInteractionId"
          },
          {
            "$ref": "#/components/parameters/xCustomerUserAgent"
          },
          {
            "$ref": "#/components/parameters/Page"
          },
          {
            "$ref": "#/components/parameters/PageSize"
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/200ConsentsConsentIdReadExtensions"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          },
          "529": {
            "$ref": "#/components/responses/SiteIsOverloaded"
          },
          "default": {
            "description": "Erro inesperado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2Security": [
              "consents"
            ]
          }
        ]
      }
    },
    "/consents/{consentId}/extends": {
      "post": {
        "tags": [
          "Consents"
        ],
        "summary": "Renovar consentimento identificado por consentId.",
        "operationId": "consentsPostConsentsConsentIdExtends",
        "description": "Método utilizado para renovação de consentimento do cliente. O consentimento só pode ser renovado caso esteja ativo (status AUTHORISED) e tenha alçada simples de aprovação (não dependa de múltiplos aprovadores).  A alteração de data efetuada pela renovação deve ser refletida também na consulta do método _GET/consents/{consentId}_.",
        "parameters": [
          {
            "$ref": "#/components/parameters/ConsentId"
          },
          {
            "$ref": "#/components/parameters/AuthorizationExtensions"
          },
          {
            "$ref": "#/components/parameters/xFapiAuthDate"
          },
          {
            "$ref": "#/components/parameters/xFapiCustomerIpAddressExtensions"
          },
          {
            "$ref": "#/components/parameters/xFapiInteractionId"
          },
          {
            "$ref": "#/components/parameters/xCustomerUserAgentExtensions"
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateConsentExtensions"
              }
            }
          },
          "description": "Payload para renovação do consentimento.",
          "required": true
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/201ConsentsCreatedExtensions"
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          },
          "403": {
            "$ref": "#/components/responses/Forbidden"
          },
          "404": {
            "$ref": "#/components/responses/NotFound"
          },
          "405": {
            "$ref": "#/components/responses/MethodNotAllowed"
          },
          "406": {
            "$ref": "#/components/responses/NotAcceptable"
          },
          "415": {
            "$ref": "#/components/responses/UnsupportedMediaType"
          },
          "422": {
            "$ref": "#/components/responses/UnprocessableEntityConsents"
          },
          "429": {
            "$ref": "#/components/responses/TooManyRequests"
          },
          "500": {
            "$ref": "#/components/responses/InternalServerError"
          },
          "504": {
            "$ref": "#/components/responses/GatewayTimeout"
          },
          "529": {
            "$ref": "#/components/responses/SiteIsOverloaded"
          },
          "default": {
            "description": "Erro inesperado.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResponseError"
                }
              }
            }
          }
        },
        "security": [
          {
            "OAuth2AuthorizationCode": [
              "openid",
              "consent:consentId"
            ]
          }
        ]
      }
    }
  },
  "components": {
    "schemas": {
...
      "CreateConsent": {
        "type": "object",
        "required": [
          "data"
        ],
        "properties": {
          "data": {
            "type": "object",
            "required": [
              "permissions",
              "loggedUser"
            ],
            "x-regulatory-required": [
              "permissions",
              "expirationDateTime"
            ],
            "properties": {
              "loggedUser": {
                "$ref": "#/components/schemas/LoggedUser"
              },
              "businessEntity": {
                "$ref": "#/components/schemas/BusinessEntity"
              },
              "permissions": {
                "type": "array",
                "items": {
                  "description": "Especifica os tipos de permissões de acesso às APIs no escopo do Open Finance Brasil - Dados cadastrais e transacionais, de acordo com os blocos de consentimento fornecidos pelo usuário e necessários ao acesso a cada endpoint das APIs. Esse array não deve ter duplicidade de itens.",
                  "type": "string",
                  "enum": [
                    "ACCOUNTS_READ",
                    "ACCOUNTS_BALANCES_READ",
                    "ACCOUNTS_TRANSACTIONS_READ",
                    "ACCOUNTS_OVERDRAFT_LIMITS_READ",
                    "CREDIT_CARDS_ACCOUNTS_READ",
                    "CREDIT_CARDS_ACCOUNTS_BILLS_READ",
                    "CREDIT_CARDS_ACCOUNTS_BILLS_TRANSACTIONS_READ",
                    "CREDIT_CARDS_ACCOUNTS_LIMITS_READ",
                    "CREDIT_CARDS_ACCOUNTS_TRANSACTIONS_READ",
                    "CUSTOMERS_PERSONAL_IDENTIFICATIONS_READ",
                    "CUSTOMERS_PERSONAL_ADITTIONALINFO_READ",
                    "CUSTOMERS_BUSINESS_IDENTIFICATIONS_READ",
                    "CUSTOMERS_BUSINESS_ADITTIONALINFO_READ",
                    "FINANCINGS_READ",
                    "FINANCINGS_SCHEDULED_INSTALMENTS_READ",
                    "FINANCINGS_PAYMENTS_READ",
                    "FINANCINGS_WARRANTIES_READ",
                    "INVOICE_FINANCINGS_READ",
                    "INVOICE_FINANCINGS_SCHEDULED_INSTALMENTS_READ",
                    "INVOICE_FINANCINGS_PAYMENTS_READ",
                    "INVOICE_FINANCINGS_WARRANTIES_READ",
                    "LOANS_READ",
                    "LOANS_SCHEDULED_INSTALMENTS_READ",
                    "LOANS_PAYMENTS_READ",
                    "LOANS_WARRANTIES_READ",
                    "UNARRANGED_ACCOUNTS_OVERDRAFT_READ",
                    "UNARRANGED_ACCOUNTS_OVERDRAFT_SCHEDULED_INSTALMENTS_READ",
                    "UNARRANGED_ACCOUNTS_OVERDRAFT_PAYMENTS_READ",
                    "UNARRANGED_ACCOUNTS_OVERDRAFT_WARRANTIES_READ",
                    "RESOURCES_READ",
                    "BANK_FIXED_INCOMES_READ",
                    "CREDIT_FIXED_INCOMES_READ",
                    "FUNDS_READ",
                    "VARIABLE_INCOMES_READ",
                    "TREASURE_TITLES_READ",
                    "EXCHANGES_READ"
                  ]
                },
                "minItems": 1,
                "example": [
                  "ACCOUNTS_READ",
                  "ACCOUNTS_OVERDRAFT_LIMITS_READ",
                  "RESOURCES_READ"
                ]
              },
              "expirationDateTime": {
                "description": "Data e hora de expiração da permissão. Reflete a data limite de validade do consentimento. Uma string com data e hora conforme especificação RFC-3339, sempre com a utilização de timezone UTC (UTC time format). \n\n[Restrição] De preenchimento obrigatório nos casos em que houver validade determinada. Em casos de consentimento com prazo indeterminado o campo não deve ser enviado.\n",
                "type": "string",
                "maxLength": 20,
                "format": "date-time",
                "pattern": "^(\\d{4})-(1[0-2]|0?[1-9])-(3[01]|[12][0-9]|0?[1-9])T(?:[01]\\d|2[0123]):(?:[012345]\\d):(?:[012345]\\d)Z$",
                "example": "2021-05-21T08:30:00Z"
              }
            }
          }
        }
      },
...
      "ResponseError": {
        "type": "object",
        "required": [
          "errors"
        ],
        "properties": {
          "errors": {
            "type": "array",
            "minItems": 1,
            "maxItems": 13,
            "items": {
              "type": "object",
              "required": [
                "code",
                "title",
                "detail"
              ],
              "properties": {
                "code": {
                  "description": "Código de erro específico do endpoint",
                  "type": "string",
                  "pattern": "[\\w\\W\\s]*",
                  "maxLength": 255
                },
                "title": {
                  "description": "Título legível por humanos deste erro específico",
                  "type": "string",
                  "pattern": "[\\w\\W\\s]*",
                  "maxLength": 255
                },
                "detail": {
                  "description": "Descrição legível por humanos deste erro específico",
                  "type": "string",
                  "pattern": "[\\w\\W\\s]*",
                  "maxLength": 2048
                }
              }
            }
          },
          "meta": {
            "$ref": "#/components/schemas/MetaError"
          }
        }
      },
...
    },
    "parameters": {
      "ConsentId": {
        "name": "consentId",
        "in": "path",
        "description": "O consentId é o identificador único do consentimento e deverá ser um URN - Uniform Resource Name.  \nUm URN, conforme definido na [RFC8141](https://tools.ietf.org/html/rfc8141) é um Uniform Resource \nIdentifier - URI - que é atribuído sob o URI scheme \"urn\" e um namespace URN específico, com a intenção de que o URN \nseja um identificador de recurso persistente e independente da localização.  \nConsiderando a string urn:bancoex:C1DD33123 como exemplo para consentId temos:\n- o namespace(urn)\n- o identificador associado ao namespace da instituição transnmissora (bancoex)\n- o identificador específico dentro do namespace (C1DD33123).\nInformações mais detalhadas sobre a construção de namespaces devem ser consultadas na [RFC8141](https://tools.ietf.org/html/rfc8141).\n",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^urn:[a-zA-Z0-9][a-zA-Z0-9-]{0,31}:[a-zA-Z0-9()+,\\-.:=@;$_!*'%\\/?#]+$",
          "minLength": 6,
          "maxLength": 256
        }
      },
      "Authorization": {
        "name": "Authorization",
        "in": "header",
        "description": "Cabeçalho HTTP padrão. Permite que as credenciais sejam fornecidas dependendo do tipo de recurso solicitado",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "[\\w\\W\\s]*",
          "minLength": 1,
          "maxLength": 2048
        }
      },
...
      "xCustomerUserAgent": {
        "name": "x-customer-user-agent",
        "in": "header",
        "description": "Indica o user-agent que o usuário utiliza.",
        "required": false,
        "schema": {
          "type": "string",
          "pattern": "[\\w\\W\\s]*",
          "minLength": 1,
          "maxLength": 100
        }
      },
...
      "xFapiAuthDate": {
        "name": "x-fapi-auth-date",
        "in": "header",
        "description": "Data em que o usuário logou pela última vez com o receptor. Representada de acordo com a [RFC7231](https://tools.ietf.org/html/rfc7231).Exemplo: Sun, 10 Sep 2017 19:43:31 UTC",
        "required": false,
        "schema": {
          "type": "string",
          "pattern": "^(Mon|Tue|Wed|Thu|Fri|Sat|Sun), \\d{2} (Jan|Feb|Mar|Apr|May|Jun|Jul|Aug|Sep|Oct|Nov|Dec) \\d{4} \\d{2}:\\d{2}:\\d{2} (GMT|UTC)$",
          "minLength": 29,
          "maxLength": 29
        }
      },
      "xFapiCustomerIpAddress": {
        "name": "x-fapi-customer-ip-address",
        "in": "header",
        "description": "O endereço IP do usuário se estiver atualmente logado com o receptor.",
        "required": false,
        "schema": {
          "type": "string",
          "pattern": "[\\w\\W\\s]*",
          "minLength": 1,
          "maxLength": 100
        }
      },
...
      "xFapiInteractionId": {
        "name": "x-fapi-interaction-id",
        "in": "header",
        "description": "Um UUID RFC4122 usado como um ID de correlação entre request e response. Campo de geração e envio obrigatório pela receptora (client) e o seu valor deve ser “espelhado” pela transmissora (server) no cabeçalho de resposta. Caso não seja recebido ou se for recebido um valor inválido, a transmissora deve gerar um x-fapi-interaction-id e retorná-lo na resposta com o HTTP Status Code 400. A receptora deve acatar o valor recebido da transmissora.",
        "required": true,
        "schema": {
          "type": "string",
          "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
          "minLength": 1,
          "maxLength": 36,
          "format": "uuid",
          "example": "d78fc4e5-37ca-4da3-adf2-9b082bf92280"
        }
      }
    },
    "securitySchemes": {
      "OAuth2Security": {
        "type": "oauth2",
        "flows": {
          "clientCredentials": {
            "tokenUrl": "https://authserver.example/token",
            "scopes": {
              "consents": "Criação do consentimento."
            }
          }
        }
      },
      "OAuth2AuthorizationCode": {
        "type": "oauth2",
        "description": "Fluxo OAuthnecessário para que a receptora tenha acesso aos endpoints de consentimento na instituição transmissora. Requer o processo de redirecionamento e autenticação do usuário.",
        "flows": {
          "authorizationCode": {
            "authorizationUrl": "https://authserver.example/code",
            "tokenUrl": "https://authserver.example/token",
            "scopes": {
              "openid": "Indica que a autorização está sendo realizada utilizando o protocolo definido pela openid.",
              "consent:consentId": "Escopo contendo o identificador único do consentimento criado para compartilhamento de dados solicitada."
            }
          }
        }
      }
    },
    "responses": {
      "UnsupportedMediaType": {
        "description": "O formato do payload não é um formato suportado.",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "BadRequest": {
        "description": "A requisição foi malformada, omitindo atributos obrigatórios, seja no payload ou através de atributos na URL.",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "Forbidden": {
        "description": "O token tem escopo incorreto ou uma política de segurança foi violada",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "GatewayTimeout": {
        "description": "GATEWAY TIMEOUT - A requisição não foi atendida dentro do tempo limite estabelecido",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "InternalServerError": {
        "description": "Ocorreu um erro no gateway da API ou no microsserviço",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "MethodNotAllowed": {
        "description": "O consumidor tentou acessar o recurso com um método não suportado",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "NotAcceptable": {
        "description": "A solicitação continha um cabeçalho Accept diferente dos tipos de mídia permitidos ou um conjunto de caracteres diferente de UTF-8",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "NotFound": {
        "description": "O recurso solicitado não existe ou não foi implementado",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "TooManyRequests": {
        "description": "A operação foi recusada, pois muitas solicitações foram feitas dentro de um determinado período ou o limite global de requisições concorrentes foi atingido",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "UnprocessableEntity": {
        "description": "A sintaxe da requisição esta correta, mas não foi possível processar as instruções presentes.",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseErrorUnprocessableEntity"
            }
          }
        }
      },
      "UnprocessableEntityDelete": {
        "description": "A sintaxe da requisição esta correta, mas não foi possível processar as instruções presentes.",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseErrorUnprocessableEntityDelete"
            }
          }
        }
      },
...
      "Unauthorized": {
        "description": "Cabeçalho de autenticação ausente/inválido ou token inválido",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "$ref": "#/components/schemas/ResponseError"
            }
          }
        }
      },
      "SiteIsOverloaded": {
        "description": "O site está sobrecarregado e a operação foi recusada, pois foi atingido o limite máximo de TPS global, neste momento.",
        "content": {
          "application/json; charset=utf-8": {
            "schema": {
              "type": "object",
              "required": [
                "errors"
              ],
              "properties": {
                "errors": {
                  "type": "array",
                  "minItems": 1,
                  "maxItems": 13,
                  "items": {
                    "type": "object",
                    "required": [
                      "code",
                      "title",
                      "detail"
                    ],
                    "properties": {
                      "code": {
                        "description": "Código de erro específico do endpoint",
                        "type": "string",
                        "pattern": "[\\w\\W\\s]*",
                        "maxLength": 255
                      },
                      "title": {
                        "description": "Título legível por humanos deste erro específico",
                        "type": "string",
                        "pattern": "[\\w\\W\\s]*",
                        "maxLength": 255
                      },
                      "detail": {
                        "description": "Descrição legível por humanos deste erro específico",
                        "type": "string",
                        "pattern": "[\\w\\W\\s]*",
                        "maxLength": 2048
                      }
                    },
                    "additionalProperties": false
                  }
                },
                "meta": {
                  "type": "object",
                  "description": "Meta informações referente a API requisitada.",
                  "required": [
                    "requestDateTime"
                  ],
                  "properties": {
                    "requestDateTime": {
                      "description": "Data e hora da consulta, conforme especificação RFC-3339, formato UTC.",
                      "type": "string",
                      "maxLength": 20,
                      "format": "date-time",
                      "example": "2021-05-21T08:30:00Z"
                    }
                  },
                  "additionalProperties": false
                }
              },
              "additionalProperties": false
            }
          }
        }
      },
      "201ConsentsCreated": {
        "description": "Consentimento criado com sucesso.",
        "headers": {
          "x-fapi-interaction-id": {
            "description": "Um UUID RFC4122 usado como um ID de correlação entre request e response. Campo de geração e envio obrigatório pela receptora (client) e o seu valor deve ser “espelhado” pela transmissora (server) no cabeçalho de resposta. Caso não seja recebido ou se for recebido um valor inválido, a transmissora deve gerar um x-fapi-interaction-id e retorná-lo na resposta com o HTTP Status Code 400. A receptora deve acatar o valor recebido da transmissora.         \n",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "minLength": 1,
              "maxLength": 36,
              "example": "d78fc4e5-37ca-4da3-adf2-9b082bf92280",
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ResponseConsent"
            }
          }
        }
      },
...
      "200ConsentsConsentIdRead": {
        "description": "Consentimento consultado com sucesso.",
        "headers": {
          "x-fapi-interaction-id": {
            "description": "Um UUID RFC4122 usado como um ID de correlação entre request e response. Campo de geração e envio obrigatório pela receptora (client) e o seu valor deve ser “espelhado” pela transmissora (server) no cabeçalho de resposta. Caso não seja recebido ou se for recebido um valor inválido, a transmissora deve gerar um x-fapi-interaction-id e retorná-lo na resposta com o HTTP Status Code 400. A receptora deve acatar o valor recebido da transmissora.        \n",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "minLength": 1,
              "maxLength": 36,
              "example": "d78fc4e5-37ca-4da3-adf2-9b082bf92280",
              "format": "uuid"
            }
          }
        },
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ResponseConsentRead"
            }
          }
        }
      },
...
      "204ConsentsConsentIdDeleted": {
        "description": "Consentimento revogado com sucesso.",
        "headers": {
          "x-fapi-interaction-id": {
            "description": "Um UUID RFC4122 usado como um ID de correlação entre request e response. Campo de geração e envio obrigatório pela receptora (client) e o seu valor deve ser “espelhado” pela transmissora (server) no cabeçalho de resposta. Caso não seja recebido ou se for recebido um valor inválido, a transmissora deve gerar um x-fapi-interaction-id e retorná-lo na resposta com o HTTP Status Code 400. A receptora deve acatar o valor recebido da transmissora.        \n",
            "schema": {
              "type": "string",
              "pattern": "^[0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{4}-[0-9a-fA-F]{12}$",
              "minLength": 1,
              "maxLength": 36,
              "example": "d78fc4e5-37ca-4da3-adf2-9b082bf92280",
              "format": "uuid"
            }
          }
        }
      }
    }
  }
}